+++ R/shapeConstr.factor.R
@@ -28,8 +28,7 @@ shapeConstr.factor <- function(x, shape, intercept = FALSE, ...) {
   }
 
   # Remove redundant constraints
-  chkc <- checkCmat(Cmat)
-  if (length(chkc$redundant) > 0) Cmat <- Cmat[-chkc$redundant, , drop = F]
+  Cmat <- Cmat[!checkCmat(Cmat)$redundant,, drop = FALSE]
 
   # Add bound attributes
   attr(Cmat, "lb") <- rep(0, NROW(Cmat))