+++ NEWS.md
@@ -14,7 +14,7 @@
 - Added the argument `complete` in inferential functions, to allow keeping or discarding aliased coefficients. Same interpretation as in `vcov.lm`.
 - `vcov` now allows returning the usual variance-covariance returned by `vcov.glm` when `trunc = FALSE`.
 - `simu_coef` renamed as `simulCoef` and now includes an argument to set the seed. 
-- `check_cmat` has been renamed `checkCmat`.
+- `check_cmat` has been renamed `checkCmat`. It also now returns logical vectors instead of vectors of indices
 - Now the `Cmat`, `lb` and `ub` used are not returned in the `control` object from the result of `glm` with `cirls.fit`.
 - Now `lb` and `ub` can be passed by term.
 - The element `aic` of a fitted `cirls` object is penalised by the number of active constraints.